agents/knowledge_graph_agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
neo4j==5.14.1
redis==5.0.1
pyyaml==6.0.1
structlog==23.2.0
prometheus-client==0.19.0
tenacity==8.2.3
httpx==0.25.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
testcontainers==3.7.1
